backend-python/pyproject.toml
[project]
name = "route-optimizer-backend"
version = "1.0.0"
description = "Route Optimizer API - Python Backend"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0,<2.4",
    "numpy>=1.24,<1.27",
    "scipy>=1.10,<1.17",
    "scikit-learn>=1.3,<1.8",
    "geopy>=2.4,<2.5",
    "urllib3>=2.0,<3.0",
    "openpyxl>=3.0,<3.2",
    "xlrd>=2.0,<2.1",
    "ortools>=9.8,<9.9",
    "boto3>=1.28",
    "flask>=3.0",
    "flask-cors>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
